refactor(api): swap queries struct for interface
- Setup sqlc to generate an interface
- Add wrapper interface to let us use the `WithTx` method
  - [Stole it from here](https://github.com/sqlc-dev/sqlc/issues/383#issuecomment-613083090)
- Update existing services and such to use new interface

+++ api/db/list_recipes.sql.go
@@ -86,7 +86,7 @@ type ListRecipesParams struct {
 	SortDir       bool
 	PublishCursor pgtype.Timestamp
 	SlugCursor    pgtype.Text
-	Limit         int32
+	Limit         int64
 }
 
 func (q *Queries) ListRecipes(ctx context.Context, arg ListRecipesParams) ([]Recipe, error) {
